Switched to Xarial.Xtoolkit
Separated View Model classes to different files
