--- xport/Themes/Generic.xaml
@@ -1,30 +1,0 @@
-<ResourceDictionary
-    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-    xmlns:local="clr-namespace:Xarial.XTools.Xport.UI">
-    
-    <local:EnumValueToHeaderConverter x:Key="valToHeaderConv"/>
-    <local:EnumItemTypeToForegroundConverter x:Key="itemFrgColorConv"/>
-
-    <Style TargetType="{x:Type local:EnumComboBox}" BasedOn="{StaticResource {x:Type ComboBox}}">
-        <Setter Property="ItemTemplateSelector">
-            <Setter.Value>
-                <local:EnumComboBoxItemTemplateSelector>
-                    <local:EnumComboBoxItemTemplateSelector.Header>
-                        <DataTemplate>
-                            <TextBlock Text="{Binding Path=Value, RelativeSource={RelativeSource AncestorType={x:Type local:EnumComboBox}}, Converter={StaticResource valToHeaderConv}}"/>
-                        </DataTemplate>
-                    </local:EnumComboBoxItemTemplateSelector.Header>
-                    <local:EnumComboBoxItemTemplateSelector.Item>
-                        <DataTemplate>
-                            <CheckBox IsChecked="{Binding Path=IsSelected}">
-                                <TextBlock Text="{Binding Path=Title}" ToolTip="{Binding Path=Description}"
-                                           Foreground="{Binding Path=Type, Converter={StaticResource itemFrgColorConv}}"/>
-                            </CheckBox>
-                        </DataTemplate>
-                    </local:EnumComboBoxItemTemplateSelector.Item>
-                </local:EnumComboBoxItemTemplateSelector>
-            </Setter.Value>
-        </Setter>
-    </Style>
-</ResourceDictionary>
