--- xport/UI/BooleanUniversalConverter.cs
@@ -1,41 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Globalization;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using System.Windows.Data;
-
-namespace Xarial.XTools.Xport.UI
-{
-    [ValueConversion(typeof(bool), typeof(object))]
-    public class BooleanUniversalConverter : IValueConverter
-    {
-        public object TrueValue { get; set; }
-        public object FalseValue { get; set; }
-
-        public object Convert(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            if (value is bool)
-            {
-                if ((bool)value)
-                {
-                    return TrueValue;
-                }
-                else
-                {
-                    return FalseValue;
-                }
-            }
-            else
-            {
-                return null;
-            }
-        }
-
-        public object ConvertBack(object value, Type targetType, object parameter, CultureInfo culture)
-        {
-            throw new NotImplementedException();
-        }
-    }
-}
