+++ xport/Xport.csproj
@@ -48,6 +48,9 @@
     <Reference Include="CommandLine, Version=2.7.82.0, Culture=neutral, PublicKeyToken=5a870481e358d379, processorArchitecture=MSIL">
       <HintPath>..\packages\CommandLineParser.2.7.82\lib\net40\CommandLine.dll</HintPath>
     </Reference>
+    <Reference Include="Newtonsoft.Json, Version=12.0.0.0, Culture=neutral, PublicKeyToken=30ad4fe6b2a6aeed, processorArchitecture=MSIL">
+      <HintPath>..\packages\Newtonsoft.Json.12.0.3\lib\net45\Newtonsoft.Json.dll</HintPath>
+    </Reference>
     <Reference Include="System" />
     <Reference Include="System.Data" />
     <Reference Include="System.Windows.Forms" />
@@ -62,18 +65,24 @@
     <Reference Include="WindowsBase" />
     <Reference Include="PresentationCore" />
     <Reference Include="PresentationFramework" />
+    <Reference Include="Xarial.XToolkit, Version=0.1.0.0, Culture=neutral, PublicKeyToken=60dcaf351d4060db, processorArchitecture=MSIL">
+      <HintPath>..\packages\Xarial.XToolkit.0.1.0\lib\net461\Xarial.XToolkit.dll</HintPath>
+    </Reference>
+    <Reference Include="Xarial.XToolkit.Wpf, Version=0.1.0.0, Culture=neutral, PublicKeyToken=60dcaf351d4060db, processorArchitecture=MSIL">
+      <HintPath>..\packages\Xarial.XToolkit.Wpf.0.1.0\lib\net461\Xarial.XToolkit.Wpf.dll</HintPath>
+    </Reference>
   </ItemGroup>
   <ItemGroup>
     <ApplicationDefinition Include="App.xaml">
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </ApplicationDefinition>
-    <Compile Include="Reflection\EnumHelper.cs" />
-    <Compile Include="UI\BooleanUniversalConverter.cs" />
-    <Compile Include="UI\EnumComboBox.cs" />
-    <Compile Include="UI\FsoBrowser.cs" />
-    <Compile Include="UI\RelayCommand.cs" />
+    <Compile Include="ViewModels\EnumDisplayNameAttribute.cs" />
     <Compile Include="ViewModels\ExporterSettingsVM.cs" />
+    <Compile Include="ViewModels\FormatExtensionAttribute.cs" />
+    <Compile Include="ViewModels\Format_e.cs" />
+    <Compile Include="ViewModels\LogWriter.cs" />
+    <Compile Include="ViewModels\ProgressHandler.cs" />
     <Compile Include="Views\ExportSettingsView.xaml.cs">
       <DependentUpon>ExportSettingsView.xaml</DependentUpon>
     </Compile>
@@ -92,10 +101,6 @@
       <DependentUpon>MainWindow.xaml</DependentUpon>
       <SubType>Code</SubType>
     </Compile>
-    <Page Include="Themes\Generic.xaml">
-      <Generator>MSBuild:Compile</Generator>
-      <SubType>Designer</SubType>
-    </Page>
     <Page Include="Views\ExportSettingsView.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
