+++ xport/packages.config
@@ -1,4 +1,7 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="CommandLineParser" version="2.7.82" targetFramework="net40" requireReinstallation="true" />
+  <package id="Newtonsoft.Json" version="12.0.3" targetFramework="net461" />
+  <package id="Xarial.XToolkit" version="0.1.0" targetFramework="net461" />
+  <package id="Xarial.XToolkit.Wpf" version="0.1.0" targetFramework="net461" />
 </packages>
